refactor(client): extract attempt() and AttemptOutcome into attempt module
Move the single-attempt connect/rpc/evict path and its outcome enum out of
the retry loop module. All seven counter emissions move verbatim with their
code paths. Doc links to crate::retry::attempt updated to crate::attempt::attempt.
